requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
cloudpickle==0.5.5
colorama==0.3.9
docutils==0.14
Flask==1.0.2
flask-marshmallow==0.10.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
future==0.17.1
face_recognition==1.2.3
google-api-core==1.5.2
google-auth==1.6.1
gunicorn==19.9.0
imageio==2.5.0
imagesize==1.1.0
iterable==0.0.1
Jinja2==2.10
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
matplotlib==2.2.3
numpy==1.14.5
opencv-python==3.4.3.18
pandas==0.23.4
path.py==11.1.0
pathlib==1.0.1
pickleshare==0.7.4
Pillow==5.2.0
requests==2.9.1
requests-file==1.4.3
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
SQLAlchemy==1.3.2
urllib3==1.23
utils==0.9.0
Werkzeug==0.14.1
